Fix mt7986 SD partitioning
The FIP partition has the EFI Boot GUID type being applied to it.  I
don't know why I thought that was nessessary for BL2 to find FIP, but it
doesn't; checking the ATF source shows FIP_BASE set as 0x680000, this is
6815744, which is the exact location in the partition table.  This is a
problem, because on updating uboot, it's attempting to load extlinux
from partition 4 not 5.