Update uboot to 2023.07-rc3
This doesn't provide additional functionality, but as it required
configuration changes to get working, I'd rather get that documented
rather than leaving this in a state it will break in the future.
I was atempting to see if updating uboot would effect the usibility of
the NAND from linux, as it doesn't work for me.