Switch uboot boot to use bootflow
If I understand correctly, this should replace
include/config_distro_bootcmd.h and make it possible to configure the
boot targets and methods via the device-tree, which means we can produce
a device-tree overlay we include in our build instead of patching the
board configuration.

This should simplify maintaining a working u-boot.