Create a full-featured kernel by default
linuxPackages_bpir3_minimal exists if the distro-standard style kernel
with mediatek options built-in.

This larger kernel has less suprises, as the minimal kernel is all about
compile time and not compatability.

Compiling this full-featured kernel takes about 90 minutes on a RK3588s
with 16GB of RAM, /tmp as a 32GB tmpfs, 8GB of zram SWAP, and 16GB SWAP
on an eMMC.  ZRAM was hit, but SWAP was not.

The same hardware compiles the minimal kernel in under 15 mintues.

If you're not actively testing patches in which case you're bottlenecked
by compiler throughput you should probably take the full kernel as it
won't need to build that often, and as such if it takes a while to
build, then it takes a while to build.