Remove bpir3_minimal kernel
This custom config is not going to be maintained going forward, so it is
getting removed.  I'm going to update the example to 6.6 shortly.

Removal is kind of gross as a flake lock update will break users
references to this, but that's probably better than leaving the minimal
one in the past forever.  Non minimal builds take about a day to build
directly on the BPiR3 though, so other build hardware would be a good
idea.

If I could get a aarch64 build box managed by someone else, producing a
slew of kernel builds for these things would be possible.