Correct mediatek wifi module name in kconfig
The module used for wifi on the BPiR3 is MT7615E, not MT7815E.  This
is likely a fat-finger.  This wasn't noticed because MT7615E is by
default a module.  Non X86 platforms on NixOS do not consider missing
kconfig options an error, making this easy to miss.

This is explicit, because when built in trouble may occour with
firmware loading.