Actually correct wifi module name
As pointed out by ghostbuster91 on the commit that changed this to
MT7615E, the correct setting is MT7915E.

I have no idea what caused me to think MT7615E was the correct
correction.

The prior kernel was still functional on the hardware, as MT7915E was
built as a module by default.